Audio: Optimize saturation inline functions in format.h
This patch adds two included header files format_generic.h and
format_hifi3.h. The latter contains intrinsics operations
optimized versions of 16, 24, and 32 bit saturations. Generic
header contains the previous C versions those will be used
for anything else but xt-xcc HiFi3 compatible build.
