+++ src/include/sof/audio/format.h
@@ -10,6 +10,22 @@
 #ifndef __SOF_AUDIO_FORMAT_H__
 #define __SOF_AUDIO_FORMAT_H__
 
+#if defined __XCC__
+#include <xtensa/config/core-isa.h>
+#if XCHAL_HAVE_HIFI3 == 1
+#define __AUDIO_FORMAT_GENERIC__	0
+#define __AUDIO_FORMAT_HIFI3__		1
+#else
+/* Generic build for e.g. hifi2 */
+#define __AUDIO_FORMAT_GENERIC__	1
+#define __AUDIO_FORMAT_HIFI3__		0
+#endif /* !XCHAL_HAVE_HIFI3 */
+#else
+/* GCC */
+#define __AUDIO_FORMAT_GENERIC__	1
+#define __AUDIO_FORMAT_HIFI3__		0
+#endif /* !__XCC__ */
+
 #include <ipc/stream.h>
 #include <stdint.h>
 
@@ -78,6 +94,15 @@
 #define SATP_INT32(x) (((x) > INT32_MAX) ? INT32_MAX : (x))
 #define SATM_INT32(x) (((x) < INT32_MIN) ? INT32_MIN : (x))
 
+/* Inline functions */
+
+#if __AUDIO_FORMAT_GENERIC__
+#include "format_generic.h"
+#endif
+#if __AUDIO_FORMAT_HIFI3__
+#include "format_hifi3.h"
+#endif
+
 static inline int64_t q_mults_32x32(int32_t x, int32_t y, const int shift_bits)
 {
 	return ((int64_t)x * y) >> shift_bits;
@@ -96,38 +121,6 @@ static inline int32_t q_mults_16x16(int16_t x, int32_t y, const int shift_bits)
 static inline int16_t q_multsr_16x16(int16_t x, int32_t y, const int shift_bits)
 {
 	return ((((int32_t)x * y) >> (shift_bits - 1)) + 1) >> 1;
-}
-
-/* Saturation inline functions */
-
-static inline int32_t sat_int32(int64_t x)
-{
-	if (x > INT32_MAX)
-		return INT32_MAX;
-	else if (x < INT32_MIN)
-		return INT32_MIN;
-	else
-		return (int32_t)x;
-}
-
-static inline int32_t sat_int24(int32_t x)
-{
-	if (x > INT24_MAXVALUE)
-		return INT24_MAXVALUE;
-	else if (x < INT24_MINVALUE)
-		return INT24_MINVALUE;
-	else
-		return x;
-}
-
-static inline int16_t sat_int16(int32_t x)
-{
-	if (x > INT16_MAX)
-		return INT16_MAX;
-	else if (x < INT16_MIN)
-		return INT16_MIN;
-	else
-		return (int16_t)x;
 }
 
 /* Fractional multiplication with shift and saturation */
